requirements.txt
asgiref==3.7.2
Django==5.0.1
django-model-utils==4.3.1
django-notifications-hq==1.8.3
jsonfield==3.1.0
pytz==2023.3.post1
sqlparse==0.4.4
swapper==1.3.0
